pyproject.toml
[project]
name = "x-langchain-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[socks]>=0.28.1",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
[tool.ruff.format]
quote-style = "single"
indent-style = "space"
[tool.ruff.lint]
select = ["I"]
ignore = []
